Removed ping test due to high complexity of implementation. Will be done later.
